build: :art: rebuild the skill system again
