+++ logic/规则Logic.md
@@ -9,7 +9,10 @@
 
 ## 游戏简介
 - 1位监管者对抗4位求生者的非对称竞技模式
-- 略
+- [本届THUAI电子系赛道为以4名同学和1名捣蛋鬼的求学与阻挠展开的非对称竞技模式，同学需要完成足够的家庭作业和考试，相互督促以避免沉迷娱乐生活，利用道具地形躲避捣蛋鬼的各种干扰诱惑，完成学业；捣蛋鬼则要极力阻止。]
+- [我们的设计是一个非对称游戏，类似第五人格，分为学生、捣蛋鬼两个阵营。在游戏中，学生修完若干课程之后通过考试即可顺利毕业，捣蛋鬼试图干扰学生使其沉迷游戏，以致于无法修完规定课程，直至挂科、退学。]
+[对于选手来说，需要提前制定好学生的学习方案以抵御对方捣蛋鬼的干扰，类似地，也需要制定好捣蛋鬼的行动策略以影响对方学生的学习，也即每队至少要写好两份代码以执行不同阵营的不同策略。]
+[当一局比赛结束（场上的学生有且仅有两种状态：退学或毕业）时，分别记录双方总得分；之后双方换边进行下半场比赛。最终将每队的学生方、捣蛋鬼方的得分相加，比较总得分判断胜负。]
 
 ## 地图
 - 地图为矩形区域，地图上的游戏对象坐标为（x, y），且x和y均为整数。x坐标轴正方向竖直向下，
@@ -93,6 +96,7 @@
         IsLockingTheDoor = 13,
         IsRummagingInTheChest = 14,
         IsClimbingThroughWindows = 15,
+        IsUsingSpecialSkill = 16,
     }
   ~~~
 - Bgm(字典)
